Fix heap corruption provoked by GC. Got the RGB Fade sample working!!
:tada:
